pyproject.toml
[project]
name = "spimex-api"
version = "0.1.0"
description = ""
authors = [
    {name = "AleksandrZaec",email = "zaecam95@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlacodegen (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "redis (>=6.1.0,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "coverage (>=7.8.2,<8.0.0)",
    "pytest-html (>=4.1.1,<5.0.0)"
]

[tool.poetry]
packages = [
    { include = "spimex_api" }
]

[tool.poetry.group.dev.dependencies]
anyio = "^4.9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
